udm branch refactor the udm module to be entirely outside the core (esp core.ising_qubo)
