+++ qamomile/core/__init__.py
@@ -1,11 +1,7 @@
 from qamomile.core import circuit as circuit_module
 from qamomile.core.bitssample import *  # noqa
 from qamomile.core.converters import qaoa as qaoa
-from qamomile.core.ising_qubo import IsingModel, UnitDiskGraph
 
 circuit = circuit_module
 
-__all__ = [
-    "qaoa", "circuit", "BitsSample", "BitsSampleSet",
-    "IsingModel", "UnitDiskGraph"
-]
+__all__ = ["qaoa", "circuit", "BitsSample", "BitsSampleSet"]
